+++ MakiseGUI/makise.h
@@ -119,17 +119,17 @@ typedef struct _MakiseGUI
 uint32_t makise_init(MakiseGUI * gui, MakiseDriver* driver, MakiseBuffer* buffer);
 void makise_deinit(MakiseGUI* gui);
 uint8_t makise_start(MakiseGUI * gui);
-uint32_t makise_pget(MakiseBuffer *b, uint16_t x, uint16_t y);
-void makise_pset(MakiseBuffer *b, uint16_t x, uint16_t y, uint32_t c);
+uint32_t makise_pget(MakiseBuffer *b, uint32_t x, uint32_t y);
+void makise_pset(MakiseBuffer *b, uint32_t x, uint32_t y, uint32_t c);
 
 /**
  * get point without checking borders
  */
-uint32_t makise_pget_fast(MakiseBuffer *b, uint16_t x, uint16_t y);
+uint32_t makise_pget_fast(MakiseBuffer *b, uint32_t x, uint32_t y);
 /**
  * set point without checking borders
  */
-void makise_pset_fast(MakiseBuffer *b, uint16_t x, uint16_t y, uint32_t c);
+void makise_pset_fast(MakiseBuffer *b, uint32_t x, uint32_t y, uint32_t c);
 
 //if partial_render = 0, then entire buffer will be rendered, if == 1, then will be rendered only first part, if == 2 then will be rendered second part
 void makise_render(MakiseGUI *gui, uint8_t partial_render);
